+++ src/main/scala/edu/ie3/simona/ontology/messages/services/WeatherMessage.scala
@@ -7,8 +7,11 @@
 package edu.ie3.simona.ontology.messages.services
 
 import edu.ie3.simona.agent.participant.data.Data.SecondaryData
+import edu.ie3.simona.agent.participant2.ParticipantAgent
 import edu.ie3.simona.ontology.messages.services.ServiceMessage.ServiceRegistrationMessage
 import edu.ie3.util.scala.quantities.Irradiance
+import org.apache.pekko.actor.typed.ActorRef
+import org.apache.pekko.actor.{ActorRef => ClassicRef}
 import squants.{Temperature, Velocity}
 
 sealed trait WeatherMessage
@@ -21,15 +24,20 @@ sealed trait WeatherMessage
   */
 object WeatherMessage {
 
+  private[services] trait WeatherInternal extends WeatherMessage
+
   /** Indicate the [[edu.ie3.simona.service.weather.WeatherService]] that the
     * requesting agent wants to receive weather for the provided coordinates
     *
+    * @param actorRef
+    *   actor ref for the agent to be registered
     * @param latitude
     *   Latitude of the requested location
     * @param longitude
     *   Longitude of the requested location
     */
   final case class RegisterForWeatherMessage(
+      actorRef: ActorRef[ParticipantAgent.Request],
       latitude: Double,
       longitude: Double,
   ) extends WeatherMessage