+++ src/main/scala/edu/ie3/simona/sim/SimonaSim.scala
@@ -110,7 +110,7 @@ object SimonaSim {
           timeAdvancer,
           scheduler,
           primaryServiceProxy.toTyped,
-          weatherService.toTyped,
+          weatherService,
         ) ++
           gridAgents ++
           extSimulationData.extDataServices.values