+++ src/main/scala/edu/ie3/simona/sim/setup/SimonaSetup.scala
@@ -13,6 +13,7 @@ import edu.ie3.simona.agent.grid.GridAgent
 import edu.ie3.simona.event.listener.{ResultEventListener, RuntimeEventListener}
 import edu.ie3.simona.event.{ResultEvent, RuntimeEvent}
 import edu.ie3.simona.ontology.messages.SchedulerMessage
+import edu.ie3.simona.ontology.messages.services.WeatherMessage
 import edu.ie3.simona.scheduler.TimeAdvancer
 import edu.ie3.simona.scheduler.core.Core.CoreFactory
 import edu.ie3.simona.scheduler.core.RegularSchedulerCore
@@ -95,7 +96,7 @@ trait SimonaSetup {
   def weatherService(
       context: ActorContext[_],
       scheduler: ActorRef[SchedulerMessage],
-  ): ClassicRef
+  ): ActorRef[WeatherMessage]
 
   /** Loads external simulations and provides corresponding actors and init data
     *