+++ src/test/scala/edu/ie3/simona/agent/participant2/ParticipantAgentInitSpec.scala
@@ -41,6 +41,7 @@ import edu.ie3.simona.test.common.input.{LoadInputTestData, PvInputTestData}
 import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
 import edu.ie3.simona.util.TickUtil.TickLong
 import org.apache.pekko.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
+import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.actor.typed.scaladsl.adapter.TypedActorRefOps
 import squants.Each
 
@@ -325,7 +326,7 @@ class ParticipantAgentInitSpec
         val refs = ParticipantRefs(
           gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryService.ref.toClassic,
-          services = Map(ServiceType.WeatherService -> service.ref.toClassic),
+          services = Map(ServiceType.WeatherService -> service.ref),
           resultListener = Iterable(resultListener.ref),
         )
 
@@ -359,6 +360,7 @@ class ParticipantAgentInitSpec
 
         service.expectMessage(
           RegisterForWeatherMessage(
+            participantAgent,
             mockInput.getNode.getGeoPosition.getY,
             mockInput.getNode.getGeoPosition.getX,
           )
@@ -385,7 +387,7 @@ class ParticipantAgentInitSpec
         val refs = ParticipantRefs(
           gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryService.ref.toClassic,
-          services = Map(ServiceType.WeatherService -> service.ref.toClassic),
+          services = Map(ServiceType.WeatherService -> service.ref),
           resultListener = Iterable(resultListener.ref),
         )
 
@@ -443,7 +445,7 @@ class ParticipantAgentInitSpec
         val refs = ParticipantRefs(
           gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryService.ref.toClassic,
-          services = Map(ServiceType.WeatherService -> service.ref.toClassic),
+          services = Map(ServiceType.WeatherService -> service.ref),
           resultListener = Iterable(resultListener.ref),
         )
 
@@ -482,6 +484,7 @@ class ParticipantAgentInitSpec
 
         service.expectMessage(
           RegisterForWeatherMessage(
+            participantAgent,
             mockInput.getNode.getGeoPosition.getY,
             mockInput.getNode.getGeoPosition.getX,
           )
@@ -509,7 +512,7 @@ class ParticipantAgentInitSpec
         val refs = ParticipantRefs(
           gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryService.ref.toClassic,
-          services = Map(ServiceType.WeatherService -> service.ref.toClassic),
+          services = Map(ServiceType.WeatherService -> service.ref),
           resultListener = Iterable(resultListener.ref),
         )